添加手牌栏，添加角色控制，修复部分错误。

+++ mznqa/Classes/logic/gameObject/cardContainer/CardBox3.h
@@ -298,7 +298,7 @@ public:
 			return result;
 
 		int luckyDog = RandomGenerator::Instance()->getRandomNumber(0, cardIDSet.size() - 1);
-		if (0 <= luckyDog && luckyDog < cardIDSet.size())
+		if (0 <= luckyDog && luckyDog < (int)(cardIDSet.size()))
 		{
 			auto it = cardIDSet.begin();
 			for (int i = 0; i < luckyDog; ++i)
